Tempus-1073-file-upload-linking-New UI changes.
