+++ ui/src/app/asset/assets.tpl.html
@@ -55,6 +55,11 @@
                             entity-id="vm.grid.operatingItem().id.id">
             </tb-alarm-table>
         </md-tab>
+        <md-tab ng-if="!vm.grid.detailsConfig.isDetailsEditMode && vm.grid.isTenantAdmin()" md-on-select="vm.grid.triggerResize()" label="{{ 'file-upload.upload' | translate }}">
+            <tb-file-upload flex entity-type="vm.types.entityType.device"
+                            entity-id="vm.grid.operatingItem().id.id">
+            </tb-file-upload>
+        </md-tab>
         <md-tab ng-if="!vm.grid.detailsConfig.isDetailsEditMode" md-on-select="vm.grid.triggerResize()" label="{{ 'asset.events' | translate }}">
             <tb-event-table flex entity-type="vm.types.entityType.asset"
                             entity-id="vm.grid.operatingItem().id.id"
@@ -75,4 +80,5 @@
                                 audit-log-mode="{{vm.types.auditLogMode.entity}}">
             </tb-audit-log-table>
         </md-tab>
+
 </tb-grid>
